Removed Get-OSVersion / Added Get-Uptime
Removed Get-OSVersion as it was not being used and is borderline
impractical. Will revisit it's usefulness later.

Added Get-Uptime as it was a command I was running often against client
machines during troubleshooting.

Minor revision to Get-CurrentUser. As an aside, the command has proven
very useful in inventory sanity checks. Will look further into improving
it going forward.